backend/requirements.txt
# Core dependencies for Multi-Agent Campaign Generation System
langchain>=0.1.16
langchain-openai>=0.1.6
langgraph>=0.0.37
openai>=1.0.0
python-dotenv>=1.0.0
reportlab==4.0.7
requests==2.31.0

# FastAPI Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1

# Optional dependencies for enhanced features
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4